[simple-memo] refactor : Controller안의 DTO객체 개별의 자바 클래스로 생성
