+++ simplememo/simplememo/src/main/java/gunhee/simplememo/dto/category/CategoryNameResponse.java
@@ -1,0 +1,13 @@
+package gunhee.simplememo.dto.category;
+
+import lombok.Getter;
+
+@Getter
+public class CategoryNameResponse {
+
+    private final String name;
+
+    public CategoryNameResponse(String name) {
+        this.name = name;
+    }
+}
