+++ simplememo/simplememo/src/main/java/gunhee/simplememo/dto/memo/MemoIdResponse.java
@@ -1,0 +1,13 @@
+package gunhee.simplememo.dto.memo;
+
+import lombok.Getter;
+
+@Getter
+public class MemoIdResponse {
+
+    private final Integer memoId;
+
+    public MemoIdResponse(Integer memoId) {
+        this.memoId = memoId;
+    }
+}
